+++ examples/main/shared/editor/EditorPlugins.jsx
@@ -587,9 +587,9 @@ const config = {
     },
     {
       name: TOOLBARS.FORMATTING,
-      shouldCreate: () => ({ desktop: true }),
-      getButtons: () => ({
-        desktop: [
+      shouldCreate: () => ({ desktop: true, mobile: { android: true } }),
+      getButtons: () => {
+        const desktopButtons = [
           FORMATTING_BUTTONS.HEADINGS,
           '|',
           FORMATTING_BUTTONS.BOLD,
@@ -599,7 +599,6 @@ const config = {
           FORMATTING_BUTTONS.TEXT_HIGHLIGHT,
           FORMATTING_BUTTONS.TITLE,
           FORMATTING_BUTTONS.BLOCKQUOTE,
-          '|',
           {
             tooltipKey: 'AlignTextDropdownButton_Tooltip',
             name: 'Alignment',
@@ -627,8 +626,20 @@ const config = {
           FORMATTING_BUTTONS.LINE_SPACING,
           FORMATTING_BUTTONS.LINK,
           FORMATTING_BUTTONS.CODE_BLOCK,
-        ],
-      }),
+        ];
+
+        const mobileButtons = [
+          FORMATTING_BUTTONS.BOLD,
+          FORMATTING_BUTTONS.ITALIC,
+          FORMATTING_BUTTONS.UNDERLINE,
+        ];
+        return {
+          desktop: desktopButtons,
+          mobile: {
+            android: mobileButtons,
+          },
+        };
+      },
     },
     { name: TOOLBARS.SIDE, addPluginMenuConfig },
     { name: TOOLBARS.MOBILE, addPluginMenuConfig },